requirements.txt
airtable-python-wrapper==0.13.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
pendulum==2.1.0
python-dateutil==2.8.1
pytzdata==2019.3
requests==2.23.0
six==1.15.0
urllib3==1.25.9
